Single-letter symbols

Variables a, b,... z are given their mvp semantic meaning.

Usage

data(lettersymbols)

Details

Twenty-six variables a-z are defined as their mvp semantic equivalent:

    a <- as.mvp("a")
    ...
    z <- as.mvp("z")
  

These objects can be generated by running script inst/symb.Rmd, which includes some further discussion and technical documentation and creates file lettersymbols.rda which resides in the data/ directory.

Letters c, q, and t might pose difficulties.

Author

Robin K. S. Hankin

Examples

data(lettersymbols)
(a+b)*(a-b)
#> mvp object algebraically equal to
#> a^2 - b^2

(x + y + z)^3 - 3*(x + y + z)*(x*y + x*z + y*z) + 3*x*y*z
#> mvp object algebraically equal to
#> x^3 + y^3 + z^3