Random multivariate polynomials

Random multivariate polynomials, intended as quick “get you going” examples of mvp objects

Usage

rhmvp(n=7,size=4,pow=6,symbols=6)
rmvp(n=7,size=4,pow=6,symbols=6)
rmvpp(n=30,size=9,pow=20,symbols=15)
rmvppp(n=100,size=15,pow=99,symbols)

Arguments

n

Number of terms to generate

size

Maximum number of symbols in each term

pow

Maximum power of each symbol

symbols

Symbols to use; if numeric, interpret as the first symbols letters of the alphabet

Details

Function rhmvp() returns a random homogeneous mvp. Function rmvp() returns a possibly nonhomogenous mvp and functions rmvpp() and rmvppp() return, by default, progressively more complicated mvp objects. Function rmvppp() returns a polynomial with multi-letter variable names.

Value

Returns a multivariate polynomial, an object of class mvp

Author

Robin K. S. Hankin

Examples

rhmvp()
#> mvp object algebraically equal to
#> 5 a b e^2 f^2 + 5 a b^2 d^3 + 5 a b^4 f + 6 a c^4 d + 4 b c^3 f^2 + 4 b^2 d^2
#> e^2 + 5 b^4 c^2
rmvp()
#> mvp object algebraically equal to
#> 6 + a^3 c + 7 b c + 7 b^2 d^2 f + 7 c + 2 d e^3 + 4 d^2 e^2 f
rmvpp()
#> mvp object algebraically equal to
#> 23 + 4 a b d^2 e^4 h^5 j m o^3 + 4 a b i l^2 n^3 + 2 a d^3 e h^5 j l^3 m n + 26
#> a f h i^2 j^3 l^2 n^2 + 16 a h k^3 l^2 o + 15 a^2 f l + 20 a^3 b^3 f^2 j l n^2
#> o^3 + 2 a^3 d^2 e f^5 j^2 k^2 n + a^3 g^4 h^2 i j k^2 l n + 3 b c e^2 f^3 h m^3
#> o + 18 b d h^2 i^2 k m + 17 b d^2 e^2 f^2 m o + 2 b e^2 f j k^2 l m^2 + 14 b^2
#> c^2 f^3 g k^3 n^2 o + 24 b^2 f^4 k m^4 + 12 b^3 f h k o^4 + 9 b^4 d h i^6 j^4 k
#> o^3 + c d f g h n + 20 c g k^2 n + 12 c m + 19 c m o + 27 c^2 e^2 g h l^4 m +
#> 26 d g h j n o^2 + 5 d^2 h^2 l + 15 f^2 i^2 k^2 m^4 o^2 + 28 h i j k + 15 h m n
#> + 22 h^2 j^2 n o^3
rmvppp()
#> mvp object algebraically equal to
#> 12 ar^2 br^2 cs^4 ct^7 cy^7 dx^5 es^4 et^7 fw^4 gr^3 gx hw^3 hx^5 iw^2 + 62
#> ar^3 at^4 au^2 av^3 az^6 bs^3 bt^7 du^6 dz^4 gs^2 gu^3 gy hz^3 iz + 63 ar^3
#> av^9 bt^9 cu^6 cw^6 dr^6 ev^2 ey^8 ft^8 fw^7 gs^6 gx^9 gz^5 hr^9 ix + 66 ar^4
#> as^4 au^6 bs^6 bw cy^5 er^4 ex^6 fx^5 fy^2 gt^5 gu^4 gx^3 gz^2 hw^7 + 17 ar^4
#> at^8 br^3 ds^7 dw^5 dy^6 dz^3 ex^8 fs^6 gt^6 gv^4 ht hw^5 ir^2 iy^8 + 4 ar^4
#> au^3 av ay^2 bt^4 cs^2 du^4 es^2 ew ez^2 fu^2 hv is + 15 ar^4 ax^6 bx^9 dv^5
#> er^5 et^4 ew^5 ex^4 ez^4 fu^4 hr^6 hx^3 ir^9 is^6 iy^8 + 31 ar^5 at br^3 bw^4
#> cs^3 cy^2 dt^2 er ez^3 fz hw hy^2 iv^2 + 80 ar^6 av^7 az^12 bu^4 bv^5 by^5 cr^4
#> ct^5 cy^7 cz fv^2 fz^6 hs^4 hy^5 it + 9 ar^8 bu^6 bz^4 cs^8 cw^6 eu^5 ex^8
#> fr^10 fs^6 fw^3 fz^6 gy^8 ir^4 iu^6 iy^6 + 61 ar^11 au^5 bx^6 bz^3 cx^3 cz^6
#> dz^9 fu^3 fv^5 gu^3 gv^14 gw^6 gx^4 hx^9 is^7 + 88 as at^3 av^2 bz cw^3 dt^2
#> es^5 ez^3 fv^2 fx^4 gv^2 gz^2 hu^4 hz^2 ix + 62 as bs bt bw^5 by^2 dz^6 eu^3
#> fv^5 fz^2 gt^2 gx^4 gz^2 hr^4 hx^2 + 89 as bt ct^2 ix + 13 as cv dy + 82 as^3
#> at^4 bs^2 cs^3 ct cy^3 dv^8 dw^3 ex^5 fs^3 fx gu^3 gx^3 gz^3 iu^2 + 18 as^3
#> at^6 av bs^6 cy^3 ds^5 ev^6 ex^2 fw^8 gv^7 gw^3 hr^9 hw^9 is^9 iv^3 + 87 as^3
#> az^5 ct^4 du^7 dz^5 er^5 ew^5 ez^7 gr^7 gz^5 hu^6 hx^5 hy^3 iu^10 ix^6 + 2 as^4
#> at ax^7 br^5 cr^2 cs^9 ds^7 dz et^4 ex^2 hr^5 ht^3 hz^5 is^4 iu^11 + 15 as^4
#> ax^4 br^4 cs^2 cv^6 cw^4 et^5 eu^7 ev^8 fr^2 gx^4 gy^3 hr^5 hv^3 hz^3 + 32 as^5
#> at^7 ay^8 bw^5 by^7 cx^4 cy^4 dr^7 dv^7 dw^10 eu^5 ft^7 gu^6 hr hs^8 + 95 as^5
#> br^6 bt^5 bv^5 bw^3 cs^3 ct^5 dt^3 ex^5 fr^5 fy^4 gy^2 hy^5 is^4 iu^6 + 46 as^6
#> av^6 ax^4 ay^4 br^10 bv^7 bw^9 by^11 cx^8 du^3 dw^8 et^5 ez^4 fr^3 hz^5 + 37
#> as^6 br^2 bw^3 cs^6 cv^5 cx^10 cy^4 ds^7 ev^4 fu^4 gs^10 gz^3 ht^7 ir^6 iu^8 +
#> 56 as^7 at^7 az^6 br^7 bz^6 cu^5 cv cw^8 dw^6 ez^7 fs^6 fy^3 hr^7 is^10 iw^8 +
#> 65 as^7 az^2 bt^4 bv^4 bw^3 bx^5 cv^3 cw^5 dt^2 dw ez^3 ft^3 gs^5 gx^6 it^6 +
#> 54 as^13 au^5 aw^12 az^6 bv^8 bz^8 dt^4 dv^4 fw^4 fz^8 hw^3 hz^7 ir^5 it^5 ix^2
#> + 53 at av^2 ay az^2 bt^2 ct^3 cw^4 cy^4 dy^5 et^2 ew^2 ey^4 fs^2 fy^5 gx^2 +
#> 15 at ax cy cz^2 fz gw hv^3 hy hz^2 iw^2 + 78 at ay dt dv^2 fw gt gw it + 67 at
#> bz ct^3 cy^5 du^2 et^4 ey^3 fr^3 fu gr^3 gt^2 gz^5 hs^7 hw^3 iv^4 + 17 at ds +
#> 79 at^2 ay^5 bx^3 by^2 cs cy^3 cz^2 ds^4 gz^2 hs^4 hx^2 hz^4 ir^2 ix^3 + 86
#> at^2 bs^2 er^2 ez ft gt ht + 69 at^3 au^2 cr^2 cu^3 cv^3 cy^4 dw^3 dy^5 dz^2
#> es^4 eu^5 fu^2 hv^4 hy^4 iv^2 + 76 at^3 cv^3 cy^3 dv^3 dw^2 dz^3 er^5 es^8 fv^4
#> fw gt^2 gx^3 hs hw^4 ix^5 + 72 at^4 ax^3 bv^2 cw^2 cz du^3 dv^2 er^2 ev ew^2
#> ex^2 hv hy^2 + 58 at^4 ax^5 by^10 ct^5 cy^6 dv^6 er^5 et^7 ey^8 ez^5 fr^4 gu^4
#> gv^6 gy^3 ix^4 + 64 at^5 ay^7 bt^3 cw^3 cx^5 et^4 eu ev^4 ez^5 ft^6 fw^4 gt^3
#> gz^5 hs^3 hz + 38 at^6 au^9 bx^5 cr^8 cs^6 cx^7 cz^14 dr^3 dx^8 dy^6 eu^3 fx^10
#> hw^5 hy^6 it^3 + 25 at^6 au^9 bx^8 cs^5 dx^5 dz^10 ev^7 fv^5 gx^3 hs^7 hu^6
#> hv^7 iw^6 ix^3 iz^6 + 72 at^7 ax^7 bt^10 bw^8 cu^4 er^2 es^4 eu^7 ez^7 gt^5
#> hs^8 ht^7 hy^3 iw^9 ix^9 + 22 au^2 az^3 bv^4 by ct dr^4 dv^3 ew^2 ey fx gt^5
#> gx^2 hz^2 + 98 au^3 ax^6 br^3 bs^3 cw^3 ex^2 fr^8 fv^7 fx^6 gu^8 gw^5 hr^2 hs^4
#> hv^6 + 67 au^3 az^4 bz^6 cy^6 dv^7 dz^7 es^5 et^9 eu^9 ey^6 ez^4 fx^12 gw^3
#> hr^8 hx^3 + 60 au^4 bs by cr ev fu gu^3 hr ir + 99 au^6 cr^4 cx^3 cy^5 dt^9
#> dy^6 er^5 et^5 ew^7 ex^5 ft^7 gv^5 hw^6 ix^2 iz^10 + 91 au^7 bt^3 bx^6 cs^5
#> cv^7 dx^4 dy^4 ez^7 fx^6 gt^5 gv^3 gz^6 ir^5 is^5 it^7 + 25 au^8 av^6 aw^2 bt^3
#> by^7 cs^4 ct^7 cz^8 et^5 ey^6 ft^8 fw^5 gy^5 hr^11 hx^10 + 81 av aw^7 ax^6 ay
#> br^4 bw^3 ct^6 cy^2 cz^5 du^3 es^5 ev^6 ew^5 gx^5 iz^2 + 7 av^3 az^6 bu^3 bv^5
#> cv^4 dx^5 dz^6 eu^4 fr^7 fx^2 gs hw^7 ir^6 iu^4 iv^5 + 35 av^5 bx^5 cr^2 dr^5
#> ds^5 dz^4 es eu fv gr^5 gs^4 gz^3 hv^3 hx^5 iu + 62 av^6 ax^3 bt^8 cx^8 cy^8
#> dv^5 eu^3 fs^5 fw^5 fy^6 ht^4 is^3 it^4 ix^5 + 36 av^6 ax^3 cs^3 cw cz^3 dx^3
#> es^4 fs^2 fu^4 gu^2 gw^4 hr ht^2 hy^4 + 3 aw ax^3 bt bw bx^3 cu^7 cx^4 cz du^3
#> dz^3 gr^5 hv^2 hx^2 hy^2 iw + 91 aw br^2 bs bw cx cy^2 dw dy er fr^2 gy^2 hu +
#> 44 aw bv^2 by ds dx gx hs + 18 aw^3 ax^8 az^8 bw^6 by cx^5 er^6 ew^4 ez^9 fv^9
#> gz^5 hx^7 hz^3 ix^8 iz^8 + 23 aw^3 bv^7 bx^13 cv^6 cy^7 dx^8 dy^10 gs^3 gt^6
#> hr^8 hs^4 hz^3 it^3 iu^12 iw^3 + 32 aw^4 az^5 bt^3 cz^3 er^4 ev^5 ew^3 fr^5
#> fy^4 gt^3 gz^4 ht^4 hz^2 it^2 iv + 4 aw^10 ax^5 bs^6 bu^4 cs^5 cw^6 cx^5 dz^8
#> fr^3 fu^7 gt^6 gv^8 hr^10 hs^6 iu^6 + 50 ax ay^2 cv^2 cx es^4 ew^2 fx gs gw^4
#> gx^2 iv + 6 ax az by^2 cz^2 dy es fw^2 hu iv + 94 ax^2 bw^2 by bz ct dz ez fx
#> fz iz + 28 ax^3 bs^5 cw^2 cy^3 du^2 dv^2 es ey^3 ft^5 fv^7 gv^4 hy^4 iu iw^5
#> iy^4 + 23 ax^3 bu^2 bv^2 dr^7 ds dz^5 er^5 fr^2 fs^2 hx^6 hz^2 is^4 it^4 iu^7
#> iv + 63 ax^5 bu^5 bw^6 cy^5 dr^2 ds^3 dt^4 dv^3 eu^4 ey fr^5 fv^6 fz^4 gu^6
#> hu^5 + 95 ax^7 ay^6 br^2 bs^5 cu^4 dr^5 du^7 dw^6 dx^6 fx^11 gs^5 gy^4 hr^5
#> hu^5 hw^6 + 66 ax^7 bu^8 cr^5 cs^7 ct^6 cx^3 dw^4 fw^5 gr^3 hs^9 hv^4 hz^4 it^6
#> iv^7 iw^6 + 2 ay + 92 ay az^3 br^4 bt^5 dv^4 fr^5 fw fy^4 gv^2 hs^2 ht^7 hu^3
#> hx^5 ir^5 iv^5 + 92 ay ct cw^2 cy cz^3 dx^5 dz^2 fv^2 gu^2 gv^2 gx^4 ht^2 iu^5
#> ix^3 + 68 ay^2 bu bv^2 bw dt^5 es^2 et ev gs^5 hz ix^3 + 74 ay^4 az^7 bx^6 cw^4
#> cy^3 cz^6 dr^5 eu^3 ev^3 ew^2 fz^2 hx hy^4 it^5 iw^3 + 6 az cy du dv^2 dw es^2
#> et^2 gw hu^3 hx^2 is^3 iw^2 ix + 38 az^2 bt^3 bu^3 ct^2 cy^2 ds^3 dv^2 dx eu^3
#> ft^2 gu^3 gx^2 hv^4 + 96 az^2 bw^3 cu^2 dt^3 du^8 fs^4 fu^3 fv gv^2 gw^3 iv^3 +
#> 77 az^6 bz^2 cr^5 ct^5 cv^3 dt du dx^2 et^4 ex^6 fr^3 ft^3 fy^2 gu^2 gx^3 + 30
#> az^7 bu^3 cw^4 cz^6 dw^5 ev^4 fw^2 fz^4 gs^3 gt^4 gu^4 hu^2 it^6 iu^3 iz^4 + 50
#> br^2 bt^5 bv^2 cs^3 dz ez^2 fv^3 fw fy^5 gr^5 gu hs^3 ht^2 it iy + 15 br^3 ct
#> eu^2 ex fr^2 fv^3 fz^2 gr gw^2 gy^3 hr^2 hs^2 ht is + br^6 bt^6 bw^5 cs^6 ct^3
#> cu^8 cv^7 cy^7 es^5 et^5 fr^5 gs^11 ht^6 hw^5 ix^9 + 100 bs bt^3 gs hv + 3 bs^2
#> bx^6 ds dz^3 es^2 fu^7 fx^3 gs^3 ht^4 hy^2 hz^2 iv^4 ix^4 iy^7 + 51 bs^10 bu^3
#> bv^2 cu^5 cx^7 dx^3 er^3 ez^9 fx^5 fy^7 gt^3 gz^6 hw is^9 + 14 bv cr dr ey^2
#> fs^3 hs^2 hv hw^2 hz^3 ir^2 iu^2 + 22 bw cx dv er fz hv hx ir it + 95 bw^3 ct
#> ds dt dw^3 er^2 ft^2 fy^3 hs^2 hw^2 iz^4 + 42 by^4 cs^5 ct^6 cv^3 dt^7 dy er^2
#> et^4 ey^10 fs gr^4 gs^7 ht^4 hx^2 iu^7 + 7 by^5 cw^2 dr^8 dw^3 er^2 ev ex^3
#> ez^3 fu^6 gy^6 hs^4 hw^5 iv ix iz^6 + 85 bz^2 cr^4 cu^4 dt^8 ew^2 ez^5 fs^3
#> gt^7 gv^2 gw^4 gz^7 ht^5 hu^3 ir^4 iv^5 + 82 bz^7 cz^4 dw^8 dx^4 eu^3 ez^7 fu^3
#> fy^5 gt^7 gv^5 gz^8 hs^4 ht^5 it^3 iw^4 + 23 cs^2 cu fr gv^2 + 80 cs^2 cu^2
#> cv^2 cx^2 du ez fv^3 gv^2 hr hy^5 ir^2 iu iy + 42 ct^5 dy^4 er^5 es^3 eu fr^2
#> fu fw^3 fx^5 fz^3 hr^4 hy^2 iy^6 + 27 cu es fz gt^2 gu ht^2 + 67 cu^8 cw^2 cx^4
#> cy^3 cz^5 ev^2 fr^3 ft^5 gv^3 gx^8 hr^2 ht^3 hz^3 iu^2 iv^3 + 96 cx dt^3 et^3
#> eu^2 ev^3 ez fr^3 fz^2 hx^3 hy is iz^4 + 55 du dy + 53 is iy