Random multivariate polynomials

Random multivariate polynomials, intended as quick “get you going” examples of mvp objects

Usage

rhmvp(n=7,size=4,pow=6,symbols=6)
rmvp(n=7,size=4,pow=6,symbols=6)
rmvpp(n=30,size=9,pow=20,symbols=15)
rmvppp(n=100,size=15,pow=99,symbols)

Arguments

n

Number of terms to generate

size

Maximum number of symbols in each term

pow

Maximum power of each symbol

symbols

Symbols to use; if numeric, interpret as the first symbols letters of the alphabet

Details

Function rhmvp() returns a random homogeneous mvp. Function rmvp() returns a possibly nonhomogenous mvp and functions rmvpp() and rmvppp() return, by default, progressively more complicated mvp objects. Function rmvppp() returns a polynomial with multi-letter variable names.

Value

Returns a multivariate polynomial, an object of class mvp

Author

Robin K. S. Hankin

Examples

rhmvp()
#> mvp object algebraically equal to
#> 2 a c d^2 f^2 + 2 a^2 b^2 c e + 5 a^2 b^3 c + 4 a^3 b c^2 + 5 b d^3 e^2 + 3 b^2
#> d f^3 + 4 b^3 c f^2
rmvp()
#> mvp object algebraically equal to
#> 9 + 2 b c^2 + 7 d^2 e^2 + 4 e^2 f^2 + 7 f
rmvpp()
#> mvp object algebraically equal to
#> 11 + 2 a d^3 e h^5 j l^3 m n + 16 a h k^3 l^2 o + 6 a m n + a^2 c f^2 k^3 l^2
#> n^3 o + 17 a^2 e f^2 g h j^2 m^3 o^4 + 15 a^2 f l + 19 a^3 b^2 d g^3 h^2 i^3
#> j^2 l^3 n + 14 a^3 b^2 e^3 f g^2 i^3 l^2 n^3 + 20 a^3 b^3 f^2 j l n^2 o^3 + 2
#> a^3 d^2 e f^5 j^2 k^2 n + a^3 g^4 h^2 i j k^2 l n + 19 a^4 b c^2 i^2 j^4 m n +
#> 3 b c e^2 f^3 h m^3 o + 9 b c^2 e^2 j k^6 n^2 o^3 + 17 b d^2 e^2 f^2 m o + 2 b
#> e^2 f j k^2 l m^2 + 24 b^2 f^4 k m^4 + 12 b^3 f h k o^4 + 9 b^4 d h i^6 j^4 k
#> o^3 + 20 c g k^2 n + 19 c m o + 27 c^2 e^2 g h l^4 m + 12 c^2 g j l m^2 o^2 +
#> 27 d^2 e h i^4 j k l m o + 14 d^2 g i^2 l o^3 + 9 d^4 f^2 i m n^2 + 17 e g k l
#> n o + 15 h m n + 22 h^2 j^2 n o^3
rmvppp()
#> mvp object algebraically equal to
#> 84 + 81 ar^2 cw^5 dr^5 dx^2 dy^6 et ev^7 ew^6 ex fu^4 fv^3 gs^3 ht^6 hv^5 hz^5
#> + 80 ar^3 ay^2 az^4 bs^3 cr^5 cu^3 cv^3 cz^2 dt^4 du^7 fw^5 gv^4 gx iv^6 ix^4 +
#> 30 ar^4 as^3 ay^6 br^4 bv^2 cs^4 cy^4 cz^3 dr^6 dv^4 er^7 fs^3 gv^5 ht^4 is^2 +
#> 96 ar^4 az br^2 bu^3 dw gy^3 hr hs^2 ht^3 hy^3 iw^3 ix + 95 ar^4 bx^3 by^2 dy
#> fv^3 gv^3 gy gz hu^2 iv^2 iz^2 + 7 ar^6 at aw bs^6 cx^6 dv^2 fx^5 gu^8 gv^3
#> hr^3 ht hu^2 hw^3 iv^5 iz^4 + 25 ar^6 av^6 aw^3 bt^5 cw^3 dz^5 es^9 ey^6 fw^8
#> gr^10 gw^5 ht^7 is^6 it^7 iz^7 + 18 ar^8 aw^8 bt^9 cr^5 dw^5 er^8 ev^3 ew^8
#> fv^6 fx hr^9 hu^6 hv^4 ir^3 iw^7 + 33 ar^9 cr^6 cu^5 cv^5 cw^6 cy^6 cz^3 dt^4
#> dx^3 fr^7 fu^5 gt^7 gz^9 hz^4 it^6 + 99 ar^10 aw^2 by^7 ct^5 du^4 dw^3 dx^5
#> es^6 gx^6 gy^9 hu^5 hv^7 hw^5 hy^5 iv^6 + 80 as au^2 ax bt^3 ct^2 ds^2 dt^2
#> dw^2 dz^2 gs hr iu ix^5 + 14 as^2 au^2 bz^3 du ft gu hx^2 ir^3 it iv^2 iz^2 +
#> 18 as^2 br^4 bu by ct ds^2 dx^3 dy^3 gs hr ir iu^2 iy^5 + 82 as^2 bw bz^3 cr^3
#> cs^3 cw^3 dx^3 dy dz^3 ey^4 ez^3 fz^2 gt^8 gv^3 hw^5 + 7 as^4 at^5 au^6 bu^7
#> bw^2 cz dt^4 er^6 et^3 fs^3 ft^5 gr^6 gw^5 hs^4 iv^7 + 92 as^5 aw^3 bt^2 cs^2
#> ct^2 cw^4 dr^3 dv^2 dx dy ex gr^2 gw^5 iy^2 + 4 as^6 bs^7 bu^3 ct^8 cy^6 dv^6
#> dw^5 dz^5 ev^10 ew^5 fs^4 fz^6 gr^8 iu^10 iz^6 + 55 as^6 bu^5 cr^3 cw^4 dv^5
#> er^2 ew^8 fs^5 gs^3 gu^4 gw^2 hy^6 iu^5 ix^4 iy + 23 as^7 at ay^4 az^4 bu^2
#> bz^2 ew^3 fs^2 ft^2 gr^5 gu^7 gz hu^5 ir^2 iw^6 + 42 as^7 bz cu^4 cz^7 dt^3
#> dy^6 dz^5 fx^4 gx gy^7 hu^2 hx^10 hy^4 iw^2 iy^4 + 37 as^8 au^6 bs^4 cr^3 cz^10
#> dt^5 dw^10 dx^4 dz^6 ez^6 fu^2 fv^3 gz^7 ht^4 iy^7 + 2 as^11 az^4 du^2 dz^9
#> ew^7 ey ez^4 fu^5 gr gz^7 hw^2 hy^4 ir^5 iu^5 iy^3 + 32 at ay^2 bu^5 bx^4 cr^4
#> cy^3 dr^3 er^5 ev^4 fy^3 ht^5 hu^4 hv^3 ir^2 iy^4 + 6 at bv^2 dr^2 er ew fx^2
#> gx hz is + 50 at bw cv^4 cw^2 cz dt^2 dw ew ex^2 hv^2 hz^4 + 31 at^2 br dx^2
#> dz^3 eu^5 ey fu^3 fv^4 gy^2 hr^3 hu iv ix^2 + 18 at^3 az^9 bv^8 ct^7 cv^3 dx^3
#> et ey^6 ez^3 fz^6 gz^5 ht^6 hw^2 iu^9 iv^9 + 96 at^3 bs^3 bt bz^4 ct^2 cv^3
#> ds^2 er^2 fv^3 gs^8 gy^3 + 3 at^4 aw^4 ax^7 bs^7 bw^3 cz^3 fw^6 fz^2 gr^3 gz
#> hz^2 ir^2 ix^2 iy^4 + 67 at^4 bs bu^3 cr^5 cu^3 cy^2 dx^5 dy^3 ey fr gs^2 hx^3
#> hy^4 iv^3 iz^7 + 85 at^5 au^4 bz^3 cr^7 ct^2 cv^4 cy^7 ds^4 du^4 fr^2 gy^8 hr^5
#> hv^2 is^3 iy^5 + 84 at^5 bu^4 bv^6 bx^3 cr^5 cv^4 cw^5 dz^6 ft^6 gr^3 hu^3 hy^6
#> hz^2 it^6 iz^7 + 66 at^7 av^6 ay^6 bv^5 cu^3 du^5 dw^3 dy^6 dz^7 ew^7 fs^8 gv^4
#> ir^4 it^4 iz^9 + 67 at^8 au^2 aw^6 bv^9 dr^10 er^6 eu^7 fs^5 ft^7 fu^9 gv^3
#> gy^9 hr^5 hx^8 iz^5 + 22 au ay br dw fv gt hu it iw + 79 au^2 aw^3 cr^2 dr^2
#> dx^3 dz ex^5 ey^2 fw^3 fx^2 gz^4 ir^4 iw^2 iz^4 + 17 au^2 ax^8 bz^6 ct^4 cy^6
#> eu^4 ey^8 fu^3 gr^3 gv^5 gx^6 gz^7 hw^8 iv^5 iy + 60 au^2 bv cy iy + 49 au^3
#> bt^4 bu^5 cx^5 dx^4 dy^8 ey^4 fs^6 fx^7 gs^7 gt^4 hr^5 hu^4 hx^5 hy^6 + 20 au^3
#> bu bz dv er eu ew^3 fs fz^2 gu gv^3 iv^2 iz^2 + 100 au^4 br^3 bw^5 cv^4 dx^10
#> eu^4 ex^3 gs^5 gu^4 gx^8 hv^6 hz^5 is iw^6 ix^6 + 54 au^5 aw^2 ay^5 br^8 bv^4
#> er^6 es^5 ev^12 ez^13 fr^8 ft^8 gt^4 gy^4 ir^7 iv^3 + 15 au^9 ax^8 az^6 bs^4
#> eu^4 ew^6 fw^9 gt^5 hr^4 hu^5 hv^5 hw^4 hy^4 iu^6 iw^3 + 3 av cu^5 dr ds^7 dw^4
#> ev ew^3 fv fw^3 fy gr^3 gs^3 it^2 iw^2 ix^2 + 6 av^2 aw az^3 cv dx er gs gt^2
#> gv hy^2 hz^2 is^3 iw^2 + 81 av^2 aw^6 bv^2 bw^5 dr ey^10 fs^7 fx^9 fy^5 gz^8
#> hs^3 hu^5 hz^7 ix^4 iz^4 + 15 av^2 br cv dr^2 dx ew ey ir^2 it^3 ix + 12 av^2
#> bv^4 cu^3 cw dx^7 dy^7 dz^4 eu^2 fu^2 gw^5 hy^7 hz^4 iv^3 iw^5 + 74 av^3 ay^5
#> br^2 dr^6 dv^4 dx^3 er^7 ex^4 fw^6 gu^5 hs^3 ht^3 hv^2 iw ix^4 + 82 av^4 ay^3
#> bs^3 bx^5 cr^8 ct^5 cy^7 dr^4 fr^7 gv^8 gw^4 hr^7 hs^3 iy^5 iz^4 + 55 av^4 ay^3
#> bv^3 bw^4 dt^7 dv^5 dw^2 dx^5 er^8 eu^3 ey^6 fv gw^6 + 48 av^5 bw^4 by^5 cy^6
#> du ev^3 fr^5 fs fw^2 gs^2 gw^2 hr^5 hu^5 ix^5 iz^5 + 56 av^8 az^10 bx^3 bz^6
#> ds^5 dt dv^8 er^6 ey^7 ez^7 fr^6 fu^7 gv^6 hr^7 iu^7 + 88 aw bt^2 bw^4 cr^2
#> ct^2 dv^3 et^2 ey^3 ez fr gy^2 hr^3 hz^5 ir^2 is^4 + 63 aw bv^7 by^8 cr^5 cw^9
#> cz^6 ds^6 dv^6 et^9 eu^3 fy^9 gu^6 ht^2 hx^8 iu^9 + 89 aw dy^2 ez fy + 68 aw^3
#> cz^5 ex^2 fs ft^2 fv gy^5 ht hy hz^2 ir + 72 aw^5 ax^7 br^10 cw^6 cx^8 cy^8
#> ds^3 dw^4 dy^10 es^6 ew^2 fx^8 gv^9 hz^9 iv^2 + 62 aw^5 ay^4 az^3 bv^5 bx^6
#> bz^5 dw^8 dx^8 et^6 ew^3 fy^8 gt^5 hs^3 iy^4 + 76 aw^5 bt^4 bv cw^3 cy^2 dt^3
#> dx^3 ey^3 gr^3 gt^3 gv^2 hu^5 hz^8 iv^4 iz + aw^9 bu^5 cz^11 ds^8 dt^7 dx^7
#> dy^3 dz^6 fu^6 fv^5 fy^6 hy^5 hz^5 iv^5 iy^6 + 50 ax ay bt^3 bv bx^5 cs cu^5
#> dz^3 ft^2 fu^2 fy^5 gr hr^2 iy^2 iz^3 + 53 ax^3 az^7 bt^10 bv^6 cv^5 cx^4 dr^10
#> du^2 fs^4 fv fx^8 fy^5 ht^6 hw^3 + 49 ax^3 fx^2 hy + 78 ay bv cv cy ex ey gt^2
#> gy + 46 ay bx^3 cx^2 dr^2 dw^4 ew^3 fr^2 ft fx gu^5 gw^2 ht^2 iw^3 ix iy^3 + 23
#> ay^2 bu^2 bz^3 cw^2 er^2 eu gr^3 hs^2 hz iv + 4 ay^3 az^7 br^5 bu^7 cx^4 dr^9
#> dx^5 dy^12 ew^6 ft^9 fu^7 fx^4 gr^7 hr^6 hw^6 + 90 ay^6 bs^3 bw^7 dr^2 du^8
#> dw^5 dz^5 ey^4 gs^5 gu^2 gw^7 gx^8 hu^5 ir^5 ix^5 + 95 ay^6 by^3 cw^6 cz^5 dt^3
#> dv^5 er^2 ez^7 ft^4 fv^3 fw^5 fy^4 gv hr^3 is^2 + 15 az br^2 bt^3 bu^2 cu cv^2
#> cx^3 dy fu^3 hs^2 hw iu^2 iy iz^2 + 4 az bs^2 dz^2 es^3 et eu^4 ex^2 fy^4 gs^4
#> hr^2 hv hz^2 it + 90 az^4 bx^3 bz^2 ds dx^6 dz^3 er^7 ex^4 ey^5 fy gv^3 gx^3
#> hv^4 hx^8 iv + 61 az^7 bs^3 bt^5 cs^3 ct^14 cv^6 cw^4 dr^6 dw^3 es^5 eu^11 fr^3
#> fw^6 gr^9 iw^9 + 50 br bw^2 cy^2 cz^2 dy^4 ev^2 ez fs + 62 br^2 bt^5 cr^2 cw^4
#> cy^2 ez fv^5 fx^2 fy fz gr^6 hs^3 iu^4 iw^2 + 63 br^4 bt^6 bu^5 cs^6 dx^5 ew^5
#> fs^5 fv^6 gt^3 gu^2 gy^4 gz^3 hs^4 hx is^5 + 36 bs^4 bz^2 cs^2 cv^4 dr^3 dv
#> dz^3 et^6 ew^3 gw^3 hz^4 iu ix^4 iy^2 + 98 bt^7 bu^8 bw^6 cs^8 cv^5 dv^3 es^3
#> ew^6 fu^3 fz^3 hw^2 it^6 iu^2 iz^4 + 15 bu bx^2 by^6 cu^5 cx^9 cy^8 dz^5 ew^4
#> ft^5 fw^7 fz^4 hr^5 ht it^2 iy^7 + 23 bu ct^2 ds dz^2 + 91 bu^2 cx^2 dw dx^2 ev
#> fu^2 fv fz gv gx hu is + 46 bu^3 dw^8 et^6 ew^4 ex^4 ez^6 ft^7 fu^10 fv^9 fx^11
#> gs^3 gv^8 hr^4 hy^5 ir^5 + 2 bu^4 cw^2 cz^6 dt^3 dz^6 er^5 et^3 ez^5 fv^7 fw^10
#> fy^2 gt^6 gw^6 is^5 it^4 + 90 bv cw du er gt hs iw^2 ix^2 + 64 bv^4 by^6 cr^5
#> cy^3 dv^3 dw^5 ex^7 ey^5 fy^3 hr^5 hs ht^4 hy^4 ir iz^3 + 25 bv^5 by^8 cx^5
#> dr^8 dy^7 dz^4 es^8 et^6 ev^2 fx^7 fy^3 hx^6 hy^5 iu^11 iw^10 + 94 bw^5 cx^3
#> dr^5 dz^4 ew^2 ex^5 ey^3 ez^3 fu^3 fz^6 gv^7 gw^4 hs^7 iu^7 iv^7 + 32 by^7 cs^6
#> dw^4 dx^4 ex^8 ey^7 ez^5 fv^5 fx^7 gt^7 gu^7 gv^10 hs^5 iu iz^8 + 98 cr^2 cx cz
#> dw^2 dz^2 er ew ez^2 fx ht hu^2 ix^3 + 57 cv^3 cz^5 dr^2 du dv^6 er^5 ew ey^6
#> fu fy^7 fz gw^7 gx^9 hv^3 ix^3 + 79 cw dt es^2 fs gs gv hy^2 + 44 cw ev ft^2 fx
#> gw gz iz + 13 dt ez gx + 2 ex + 17 ey gz + 9 fy + 55 gs gx + 97 hr hy iu^2